\name{prepare.file}
\alias{prepare.file}
\title{Remove all the multiple blanks of the selected R output, in order to copied correctly to LibreOffice Calc.}
\usage{
prepare.file(afile)
}
\arguments{
\item{afile}{A file}
}
\value{
The same file with reduced blank spaces.
}
\description{
Remove all the multiple blanks of the selected R output, in order to copied correctly to LibreOffice Calc.
}
